chore: Add integration tests for http tracking on web
Web doesn't actually use our http tracking module, but the tests for http tracking live in that module. To ensure we don't break things in web distributed tracing, add web as an integration test platform to that package.

Additionally, add sanity tests for the Dio, which we assume works because it uses HttpOverrides on mobile (which our plugin intercepts) and XMLHttpRequest on web (which the Browser SDK intercepts).  Adding an integration test to ensure it's working on all supported platforms and that we don't break it moving forward.

refs: #663