feat: Use Flutter timestamps for RUM actions.
Instead of using the time reported by the native SDKs when we perform RUM actions, we're going to use the time reported by Dart / Flutter before it is transferred to the native platform. This should be more accurate, as it won't be slowed by the amount of time it takes the method channel to process.

refs: RUM-8246